Move validation util file into subdirectory
